\name{pd.primeview}
\Rdversion{1.1}
\docType{package}
\alias{pd.primeview}
\title{
  Annotation package for pd.primeview.
}
\description{
  Annotation package for pd.primeview built with pdInfoBuilder.
}
\details{
  This package is to be used in conjunction with the oligo package.
}
\keyword{package}
